BoltAI-Projects/requirements.txt
openai
--upgrade boltiotai
python-dotenv
python-telegram-bot
pillow
